+++ R/chromote_session.R
@@ -1,19 +1,14 @@
-# This represents one _session_ in a Chromote object. Note that in the Chrome
-# DevTools Protocol a session is a debugging interface connected to a
-# _target_; a target is a browser window/tab, or an iframe. A single target
-# can have more than one session connected to it.
-
 #' ChromoteSession class
+#'
+#' @description
+#' This represents one _session_ in a Chromote object. Note that in the Chrome
+#' DevTools Protocol a session is a debugging session connected to a _target_,
+#' which is a browser window/tab or an iframe.
+#'
+#' A single target can potentially have more than one session connected to it,
+#' but this is not currently supported by chromote.
+#'
 #' @export
-#' @param timeout_ Number of seconds for \pkg{chromote} to wait for a Chrome
-#' DevTools Protocol response. If `timeout_` is [`rlang::missing_arg()`] and
-#' `timeout` is provided, `timeout_` will be set to `2 * timeout / 1000`.
-#' @param timeout Number of milliseconds for Chrome DevTools Protocol execute a
-#' method.
-#' @param width Width, in pixels, of the `Target` to create if `targetId` is
-#'   `NULL`
-#' @param height Height, in pixels, of the `Target` to create if `targetId` is
-#'   `NULL`
 #' @param targetId
 #'   [Target](https://chromedevtools.github.io/devtools-protocol/tot/Target/)
 #'   ID of an existing target to attach to. When a `targetId` is provided, the
@@ -49,6 +44,7 @@ ChromoteSession <- R6Class(
     #'   from the parent `Chromote` object. If `TRUE`, enable automatic
     #'   event enabling/disabling; if `FALSE`, disable automatic event
     #'   enabling/disabling.
+    #' @param width,height Width and height of the new window.
     #' @param wait_ If `FALSE`, return a [promises::promise()] of a new
     #'   `ChromoteSession` object. Otherwise, block during initialization, and
     #'   return a `ChromoteSession` object directly.
@@ -76,9 +72,11 @@ ChromoteSession <- R6Class(
           wait_ = FALSE
          )$
           then(function(value) {
+            private$target_id <- value$targetId
             parent$Target$attachToTarget(value$targetId, flatten = TRUE, wait_ = FALSE)
           })
       } else {
+        private$target_id <- targetId
         p <- parent$Target$attachToTarget(targetId, flatten = TRUE, wait_ = FALSE)
       }
 
@@ -131,7 +129,7 @@ ChromoteSession <- R6Class(
         # returning p, because the call to ChromoteSession$new() always
         # returns the new object. Instead, we'll store it as
         # private$init_promise_, and the user can retrieve it with
-        # b$init_promise().
+        # b$get_init_promise().
         private$init_promise_ <- p$then(function(value) self)
       }
 
@@ -157,11 +155,9 @@ ChromoteSession <- R6Class(
     #' if (interactive()) b$view()
     #' ```
     view = function() {
-      tid <- self$Target$getTargetInfo()$targetInfo$targetId
-
       # A data frame of targets, one row per target.
       info <- fromJSON(self$parent$url("/json"))
-      path <- info$devtoolsFrontendUrl[info$id == tid]
+      path <- info$devtoolsFrontendUrl[info$id == private$target_id]
       if (length(path) == 0) {
         stop("Target info not found.")
       }
@@ -192,15 +188,12 @@ ChromoteSession <- R6Class(
         return(invisible())
       }
 
-      p <- self$Target$getTargetInfo(wait_ = FALSE)
-      p <- p$then(function(target) {
-        tid <- target$targetInfo$targetId
-        # Even if this session calls Target.closeTarget, the response from
-        # the browser is sent without a sessionId. In order to wait for the
-        # correct browser response, we need to invoke this from the parent's
-        # browser-level methods.
-        self$parent$protocol$Target$closeTarget(tid, wait_ = FALSE)
-      })
+      # Even if this session calls Target.closeTarget, the response from
+      # the browser is sent without a sessionId. In order to wait for the
+      # correct browser response, we need to invoke this from the parent's
+      # browser-level methods.
+      p <- self$parent$protocol$Target$closeTarget(private$target_id, wait_ = FALSE)
+
       p <- p$then(function(value) {
         if (isTRUE(value$success)) {
           self$mark_closed()
@@ -420,21 +413,41 @@ ChromoteSession <- R6Class(
     #' when the `ChromoteSession` has created a new session. Otherwise, block
     #' until the `ChromoteSession` has created a new session.
     new_session = function(width = 992, height = 1323, targetId = NULL, wait_ = TRUE) {
-      self$parent$new_session(width = width, height = height, targetId = targetId, wait_ = wait_)
+      create_session(
+        chromote = self$parent,
+        width = width,
+        height = height,
+        targetId = targetId,
+        wait_ = wait_
+      )
     },
 
     #' @description
     #' Retrieve the session id
-    #'
-    #' ## Examples
-    #'
-    #' ```r
-    #' b <- ChromoteSession$new()
-    #' b$get_session_id()
-    #' #> [1] "05764F1D439F4292497A21C6526575DA"
-    #' ```
     get_session_id = function() {
       private$session_id
+    },
+
+    #' @description
+    #' Create a new session that connects to the same target (i.e. page)
+    #' as this session. This is useful if the session has been closed but the target still
+    #' exists.
+    respawn = function() {
+      if (!private$is_active_) {
+        stop("Can't respawn session; target has been closed.")
+      }
+
+      create_session(
+        chromote = self$parent,
+        targetId = private$target_id,
+        auto_events = private$auto_events
+      )
+    },
+
+    #' @description
+    #' Retrieve the target id
+    get_target_id = function() {
+      private$target_id
     },
 
     #' @description
@@ -549,7 +562,7 @@ ChromoteSession <- R6Class(
     #' @description Initial promise
     #'
     #' For internal use only.
-    init_promise = function() {
+    get_init_promise = function() {
       private$init_promise_
     },
 
@@ -564,6 +577,7 @@ ChromoteSession <- R6Class(
 
   private = list(
     session_id = NULL,
+    target_id = NULL,
     is_active_ = NULL,
     event_manager = NULL,
     pixel_ratio = NULL,
@@ -576,3 +590,30 @@ ChromoteSession <- R6Class(
     }
   )
 )
+
+
+# Wrapper around ChromoteSession$new() that can return a promise
+create_session <- function(chromote = default_chromote_object(),
+                           width = 992,
+                           height = 1323,
+                           targetId = NULL,
+                           wait_ = TRUE,
+                           auto_events = NULL) {
+
+  session <- ChromoteSession$new(
+    parent = chromote,
+    width = width,
+    height = height,
+    targetId,
+    auto_events = auto_events,
+    wait_ = wait_
+  )
+
+  if (wait_) {
+    session
+  } else {
+    # ChromoteSession$new() must return a ChromoteSession object so we need a
+    # side-channel to return a promise
+    session$get_init_promise()
+  }
+}